Extracted the logo uploader into its own partial (#15705)
