+++ app/views/admin/creator_settings/_form.html.erb
@@ -7,18 +7,7 @@
   <%= text_field_tag :community_name, "", placeholder: "Climbing Life", class: "crayons-textfield", required: true %>
 </div>
 
-<div class="crayons-field mt-6 align-left">
-  <%= label_tag :logo, class: "crayons-field__label" do %>
-    Logo
-    <span class="crayons-field__required crayons-tooltip" data-tooltip="This will set the logo for your Forem" aria-describedby="logo-subtitle"></span>
-    <p id="logo-subtitle" class="crayons-field__description">You can upload a PNG or JPEG image.</p>
-  <% end %>
-  <div class="flex flex-1 gap-4">
-    <%= file_field_tag :logo, accept: @logo_allowed_types.to_s, data: { "max-file-size-mb": @max_file_size.to_s, action: "change->creator-settings#previewLogo" }, aria: { describedby: "logo-subtitle" } %>
-    <div data-creator-settings-target="previewLogo">
-    </div>
-  </div>
-</div>
+<%= render partial: "admin/shared/logo_upload", locals: { allowed_types: @logo_allowed_types, max_file_size: @max_file_size } %>
 
 <div class="crayons-field mt-6 align-left">
   <%= label_tag :primary_brand_color_hex, class: "crayons-field__label" do %>
